[refactorDayPickerInput] Refactor DayPickerInputField.
